Consolidate ETNN coordinate policy implementation
