--- configs/transforms/model_defaults/etnn_lappe.yaml
@@ -1,10 +1,0 @@
-defaults:
-  - data_manipulations@lappe_coordinates: laplacian_encodings
-  - liftings@_here_: ${get_required_lifting:${dataset},${model}}
-
-lappe_coordinates:
-  max_pe_dim: 3
-  include_eigenvalues: false
-  include_first: false
-  concat_to_x: false
-  method: exact
